requirements.txt
beautifulsoup4==4.6.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-instagram==0.3.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.1
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
lxml==4.4.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.6.1
wagtail-blocks==0.6.8
wagtailfontawesome==1.1.4
webencodings==0.5.1
Willow==1.1
